fixup: restore paidBy / standingPath; pack NodeInfo; fix quote
PR review surfaced three more fields with the same ADR-canonical-state
pattern as regionPrev:

- Incident.paidBy is documented in ADR 033 § 87 ("Governor or
  emergency-multisig that authorized") and surfaces in the canonical
  PendingClaimQueued event signature. Restored at original position,
  pre-PR storage layout preserved.
- BlacklistAppeal.standingPath is part of ADR 031's persisted appeal
  shape (§ 53, § 110, § 127, § 229) and is read on-chain by
  admissibility condition (c) of the synthetic-standing clawback
  (§ 216): `status == Open && standingPath == TokenHolder && ...`.
  Restored to the struct; param + enum-range check + event unchanged.

Pack NodeInfo by narrowing lastMultiaddrUpdate from uint256 to uint64:
ethAddress(20) + active(1) + lastMultiaddrUpdate(8) = 29B share one
slot, saving one SSTORE per _writeNodeInfo and updateMultiaddrs. Rust
alloy mirror updated in lock-step.

Fix the regionPrev docstring quote of ADR 030 § 52 — it was missing
the "entry is global" disjunct, which inverted the predicate's
meaning with the "iff" reading. Test docstring updated to note event
coverage.

Net change vs. the original PR: 3 fields removed instead of 5
(lastSettlementAt; NodeInfo.{registeredAt, firstRegisteredAt}; and
slashInflowOf — the only one with no ADR backing). Two ABI-breaking
struct edits remain: NodeInfo (smaller + narrower) and Incident
unchanged from pre-PR shape.

+++ contracts/src/CapacityBond.sol
@@ -223,13 +223,14 @@ contract CapacityBond is
 
     /// @notice Previous-region snapshot taken on each `updateRegion` call.
     ///         Pre-positioned for the ADR 030 § 52 blacklist-scope ripening
-    ///         predicate (*"a node is in scope iff entry.region == regionHint,
-    ///         OR (block.timestamp - effective < REGION_STABILITY_WINDOW AND
-    ///         entry.region == regionPrev)"*), which is an on-chain check that
-    ///         must consult this slot at scope-test / slash-eligibility time.
-    ///         No contract reads it in this revision — ContentBlacklist's
-    ///         scope test currently covers only GLOBAL ∪ region; the ripening
-    ///         leg lands with the ADR 030 enforcement PR.
+    ///         predicate (*"a node is in scope iff the entry is global, OR
+    ///         entry.region == regionHint, OR (block.timestamp - effective <
+    ///         REGION_STABILITY_WINDOW AND entry.region == regionPrev)"*),
+    ///         which is an on-chain check that must consult this slot at
+    ///         scope-test / slash-eligibility time. No contract reads it in
+    ///         this revision — ContentBlacklist's scope test currently
+    ///         covers only GLOBAL ∪ region; the ripening leg lands with the
+    ///         ADR 030 enforcement PR.
     mapping(address operator => string) public regionPrev;
 
     /// @notice Last `updateRegion` timestamp; 0 means region has never been
@@ -298,9 +299,13 @@ contract CapacityBond is
 
     struct NodeInfo {
         bytes32 nodeId;
+        // `ethAddress` (20B) + `active` (1B) + `lastMultiaddrUpdate` (8B) =
+        // 29 bytes — pack into one storage slot. Don't separate or widen
+        // any of these three without re-checking the packing or every
+        // `_writeNodeInfo` pays an extra SSTORE.
         address ethAddress;
         bool active;
-        uint256 lastMultiaddrUpdate;
+        uint64 lastMultiaddrUpdate;
         bytes multiaddrs;
         string regionHint;
     }
@@ -820,7 +825,7 @@ contract CapacityBond is
         info.nodeId = nodeId;
         info.ethAddress = msg.sender;
         info.active = true;
-        info.lastMultiaddrUpdate = block.timestamp;
+        info.lastMultiaddrUpdate = uint64(block.timestamp);
         info.multiaddrs = multiaddrs;
         info.regionHint = regionHint;
     }
@@ -900,7 +905,7 @@ contract CapacityBond is
         if (block.timestamp < readyAt) revert MultiaddrCooldownActive(readyAt);
 
         info.multiaddrs = multiaddrs;
-        info.lastMultiaddrUpdate = block.timestamp;
+        info.lastMultiaddrUpdate = uint64(block.timestamp);
 
         emit NodeMultiaddrUpdated(info.nodeId, multiaddrs);
     }
